-32 bit was not enough accuracy, and replacing the doubles with floats did not ammount to much of a peformance increase.

+++ headers/kernelDefs.h
@@ -17,22 +17,22 @@ extern __constant__ int trainingsize;
 
 
 //main kernels
-__global__ void genWeightsKern( kernelArray<float> Vec, size_t in, kernelArray<int> params, size_t offset);
+__global__ void genWeightsKern( kernelArray<double> Vec, size_t in, kernelArray<int> params, size_t offset);
 
-__global__ void NetKern(kernelArray<float> Vec, kernelArray<int> params, Order* connections, int hour, kernelArray<double> meanCh,
+__global__ void NetKern(kernelArray<double> Vec, kernelArray<int> params, Order* connections, int hour, kernelArray<double> meanCh,
                         kernelArray<double> stdCh, size_t device_offset);
 
-__global__ void reduceFirstKern(kernelArray<float> Vec,
-                                kernelArray<float> per_block_results,
+__global__ void reduceFirstKern(kernelArray<double> Vec,
+                                kernelArray<double> per_block_results,
                                 kernelArray<int> params, size_t device_offset);
 
-__global__ void reduceSecondKern(kernelArray<float> per_block_results, kernelArray<int> params, float *result);
+__global__ void reduceSecondKern(kernelArray<double> per_block_results, kernelArray<int> params, double *result);
 
-__global__ void normalizeKern(kernelArray<float> Vec, kernelArray<int> params, float *avgFitness, size_t device_offset);
+__global__ void normalizeKern(kernelArray<double> Vec, kernelArray<int> params, double *avgFitness, size_t device_offset);
 
-__global__ void evolutionKern(kernelArray<float> Vect, kernelArray<int> params, int *childOffset, int *realGridSize, size_t in, size_t device_offset);
+__global__ void evolutionKern(kernelArray<double> Vect, kernelArray<int> params, int *childOffset, int *realGridSize, size_t in, size_t device_offset);
 
-__global__ void bitonicSortKern(kernelArray<float> Vec, kernelArray<int> params, int j, int k, size_t device_offset);
+__global__ void bitonicSortKern(kernelArray<double> Vec, kernelArray<int> params, int j, int k, size_t device_offset);
 
-__global__ void cutoffKern(kernelArray<float> Vec, kernelArray<int> params, int *childOffset, int *evoGridSize, float *avgFitness, size_t device_offset);
+__global__ void cutoffKern(kernelArray<double> Vec, kernelArray<int> params, int *childOffset, int *evoGridSize, double *avgFitness, size_t device_offset);
 #endif
