+++ headers/utilFunc.h
@@ -11,9 +11,9 @@ __host__ __device__ float normalize(float x, float mean, float stdev);
 
 __host__ __device__ float shift(float x, float max, float min);
 
-__host__ __device__ float ActFunc(float x);
+__host__ __device__ double ActFunc(double x);
 
-__host__  __device__ float scoreFunc(float whenGuess, int whenAns, float latGuess, float lonGuess, float latAns, float lonAns);
+__host__  __device__ double scoreFunc(double whenGuess, int whenAns, double latGuess, double lonGuess, double latAns, double lonAns);
